requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.9
pillow==11.3.0
Pygments==2.19.2
sqlparse==0.5.3
